chore: add pagination to docs of message list endpoint

+++ server/Webby.ChatService/internal/handlers/handler.go
@@ -17,7 +17,7 @@ type chatService interface {
 
 type messageService interface {
 	SaveMessage(ctx context.Context, senderID, chatID uuid.UUID, content string) error
-	List(ctx context.Context, chatID, userID uuid.UUID, limit, page int) ([]models.RichMessage, int64, error)
+	List(ctx context.Context, chatID, userID uuid.UUID, limit, page int) ([]models.RichMessage, int, error)
 	Update(ctx context.Context, chatID, messageID, userID uuid.UUID, content string) error
 	Delete(ctx context.Context, chatID, messageID, userID uuid.UUID) error
 }
